Move matrix tests to separate file
... and re-enable parallel subtests
